requirements.txt
matplotlib==3.8.0
numpy==1.26.0
Pillow==10.0.1
